requirements.txt
dash==2.8.1
dash-bootstrap-components==1.3.1
keras==2.10.0
numpy==1.22.4
Pillow==9.4.0
tensorflow==2.10.0
ultralytics==8.0.35